Added scaling drawing procedures, added example of moving and facing mouse sprite.

+++ lib/stds/init.h
@@ -8,7 +8,7 @@ App app;
 /*
  *
  */
-extern void initGame(const char*, int, int);
+extern void initGame(const char*, uint16_t, uint16_t);
 
 /*
  *
